Move nondet handling to a new file
